Move contig_tracks from isbi_utils into utils
